Move stream to api
